refactor(wandb): move config flattening up out of infra; drop short-name threading
Relocate the typed-metric-list flattening from infra.wandb into
eval_metrics.wandb_config_dict, which builds the metric short-name map itself. This
removes the metric_short_names param that was threaded through
init_pd_run -> with_wandb -> init_wandb purely to keep infra from importing the metric
registries. infra.wandb.init_wandb now just logs a ready dict; callers render it via
wandb_config_dict. infra stays a leaf (no upward dep), and there's no param conga line.
